Refactor "active" object -> "player" object
The active object had a lot of responsibility and inconsistent naming concepts, so this commit encapsulates all interaction, active tetro, and next tetro concerns under the "player" object.

A big part of this refactor is the addition of the "tetro" prototype. Using setmetatable for class-like property inheritance, the definitions in the tetro library all now inherit functions for returning current_shape() and rotate(). This helps clean up the player code a bit.

Included with this refactor are changes to the "tetros" array, now renamed to "tetro_library".